# Passing down the App version to the help/version commands