--- a/cyclopts/core.py
+++ b/cyclopts/core.py
@@ -330,6 +330,7 @@ class App:
                 name=self._version_flags,
                 help_flags=[],
                 version_flags=[],
+                version=self.version,
                 help="Display application version.",
             )
 
@@ -347,6 +348,7 @@ class App:
                 name=self._help_flags,
                 help_flags=[],
                 version_flags=[],
+                version=self.version,
                 help="Display this message and exit.",
             )