@BrianPugh 2024-08-16
Thank you! I really need to get around to profiling real world situations like this. What % of the launch time did this take up? Is there any other low hanging optimizations?