@daudef 2024-08-16
The import time (under the profiler) of my app went from 714ms to 204ms, so quite a lot 😅. I don't think there is any other significant optimizations for the App instantiation since the time spent instantiating went from 513ms to 0.424ms. The main __call__ call is at 84ms, not really an issue in my case, and it spends 27ms  in `format_str`, 25ms in rich `render` and 24 ms in `_resolve_console` so i don't think it is easily optimisable.